Extract Panel
