+++ pages/proposals/[proposal].tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from 'next'
 import { useRouter } from 'next/router'
-import Layout from '../../components/layout'
+import { Layout } from '../../components/layout'
 import { getResult, useCardanoSerializationLib } from '../../cardano/serialization-lib'
 import { ErrorMessage, Loading } from '../../components/status'
 import { TransactionViewer } from '../../components/transaction'
